refactor: extract handleRequestError into proxy-error-handler.js
Move the request error handling function from proxy-request.js into a
dedicated proxy-error-handler.js module. This isolates error classification
and HTTP response construction from the main proxy plumbing.

- proxy-error-handler.js: handleRequestError with full JSDoc
- proxy-request.js: imports from new module, re-exports for callers
- Dockerfile: updated COPY to include new file

Closes #5634
